shio years to 2045
added shio year data through 2045; updated definition of Seconds for Windows so it's 64bit on 64bit systems

+++ py_gnome/lib_gnome/TypeDefs.h
@@ -41,7 +41,8 @@ typedef short LandType;
 #ifdef _MSC_VER
 	#if _WIN64
 		// for now we will not change the type for windows64
-		typedef long Seconds; // duration in seconds, or seconds since 1904
+		//typedef long Seconds; // duration in seconds, or seconds since 1904
+		typedef long long Seconds; // duration in seconds, or seconds since 1904
 	#else
 		#ifndef pyGNOME
 			typedef unsigned long Seconds;
